xiangqin-backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
fake-useragent==1.5.1
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
jsonpath==0.82.2
MarkupSafe==2.1.5
pillow==10.2.0
pyasn1==0.6.0
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
